Add an introduction screen
This replaces the old poor-man's help screen and hopefully serves its
purpose of telling users about the hotkeys better.

In the wake of adding this, Gromit-MPX also received a basic config
system based on GKeyFile, which some day might replace the handwritten
config parser we rely on now.

GKeyFile was not there when Simon started the original Gromit (it
appeared around 2004), so he had to resort to GScanner.

Closes #56